requirements.txt
black==21.9b0
cachelib==0.9.0
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
click==8.1.3
colorama==0.4.5
cryptography==38.0.1
Flask==2.0.1
Flask-Cors==3.0.10
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3.post0
gunicorn==20.1.0
idna==2.8
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.14.0
mypy-extensions==0.4.3
numpy==1.23.4
oauthlib==3.0.1
pathspec==0.10.1
platformdirs==2.5.2
psycopg2-binary==2.9.4
pycparser==2.21
PyMySQL==1.0.2
pyOpenSSL==22.0.0
python-dotenv==0.19.0
regex==2022.9.13
requests==2.21.0
six==1.16.0
SQLAlchemy==1.4.25
tomli==1.2.3
typing_extensions==4.4.0
urllib3==1.24.3
Werkzeug==2.0.3
PyJWT==2.6.0